Remove invalid checks from deploy precheck
The deploy-precheck script is now stored in a versioned
directory under /opt/software/rel-<ver>, so there is no
need to check if the feed directory exists.

This commit removes deprecated variables and checks that are
not needed as result of the versioning of the release scripts
directory.

Test Plan
PASS: run "software deploy precheck" successfully

Story: 2010676
Task: 49633

Change-Id: I42b00143234f4aa5a3c871bbd6900f3bfce43c8d
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>