Create deploy host alarms
This commit creates two alarms, one for the scenario when
"deploy host" executes with success, and one for the scenario
where it fails.

The success alarm will be cleared after the host is unlocked,
in [1], and the failure alarm is cleared when the "deploy host"
is retried for the same host.

[1] https://review.opendev.org/c/starlingx/config/+/916835

Test Plan
PASS: run "deploy host" successfully and verify the alarm
PASS: run "deploy host" with failure and verify the alarm

Story: 2010676
Task: 49937

Depends-on: https://review.opendev.org/c/starlingx/fault/+/916786

Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>
Change-Id: Ie1a32f725511658616f016d56de38cd04f1d9c5d