Deploy host validation blocking non RR deploy
During the deploy host if the host is not locked the validations
will block the action however the validation is wrong in case of
a non reboot required deployment, this commit fix this validation.

It also changes the logic to check if the release is major or not.

Test Plan:
PASS: Deploy host blocked to deploy a RR with host unlocked.
PASS: Deploy host not blocked to deploy a non RR with host unlocked.
PASS: Deploy major release.

Story: 2010676
Task: 49967

Change-Id: I1c76ab5a02148588421a3d33805ebe716a570b81
Signed-off-by: Luis Eduardo Bonatti <LuizEduardo.Bonatti@windriver.com>