Check to determine if a major release deploy is in progress.
In-service patch install is currently rejected expecting
the host to be locked as the framework considers it as
major release deploy. As part of this fix, check is added to
determine if the release is a major release or a patch release.

Test Plan:
PASSED: On SX system, perform in-service patch install
successfully.

Closes-Bug: 2063854

Change-Id: I7ef856c41c6721daa68310666097d65e816056b1
Signed-off-by: Vanathi.Selvaraju <vanathi.selvaraju@windriver.com>