Software deploy complete implementation
This commit implements the software deploy complete, it checks
if the deployment is as "activate-done" and all hosts are in
deployed state. If pass all validations it will move the deploy
state to 'completed'

Test Plan:
PASS: State moved to completed with host in deployed state.
PASS: State blocked to move to complete with state different than
'activate-done'
PASS: State moved to completed with patch release.
PASS: State blocked to move to complete with host not in deployed state.

Story: 2010676
Task: 50022

Change-Id: I35d54a8a9b485e90170d624d62721411ca2ca3df
Signed-off-by: Luis Eduardo Bonatti <LuizEduardo.Bonatti@windriver.com>