software deploy abort implementation
This commit implements the software deploy abort, it checks if the
deployment is in one of this states: activate, activate-done,
activate-failed, completed, host, host-done and host-failed.

Changes the order of from_release and to_release in order to
start the host-rollback so the target release will be the
from_release. It also changes the commit_id and feed_repo based on
from_release.

Set the correct state for the deploy, in case of a host-failed
a check against the sw_version will be needed to set the state as
rollback-deployed in case of the sw_version is the same as
from_release, or set rollback-pending otherwise.

Raises an exception case deployment is not major release because
the abort is not supported to patch releases.

Test Plan:

PASS: Deploy abort sets the correct data and states at
software.json after unlock of host deploy.
PASS: Deploy abort sets the correct data and states at
software.json before unlock of host deploy.
PASS: Right to_release and state with deploy failed state.

Task: 50223
Story: 2010676

Depends-On: https://review.opendev.org/c/starlingx/update/+/920763

Change-Id: Idce97e049fdadac2d3041a5f29488a9d77774493
Signed-off-by: Luis Eduardo Bonatti <LuizEduardo.Bonatti@windriver.com>