Add host rollback command
This commit adds the commands to the software client
that will be used to rollback hosts to the previous
deployment ostree commit-id. The backend logic for
this command is not covered by this commit and will
be implemented in a follow-up commit.

The commit fixes minor tox issues as well.

Test Plan
PASS: run host-rollback command and verify the output
      returned is correct

Story: 2010676
Task: 50201

Change-Id: Ib1267f493e274c7884d4e5ec3222050a96f42cc5
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>