Fix kernel agent hook
In multi-node deployment scenario, the host hieradata may not
exist yet when the agent hooks run, since the hieradata is refreshed
during the host-unlock, which occurs after deploy host command is
executed. This causes the kernel agent hook to fail in this scenario.

This commit makes the hook uses the FROM release hieradata if the
host hieradata is not available yet, and this should not have impact
since the kernel manifest does not use any host specific hieradata.

Test Plan
PASS: instantiate and run kernel hook manually in all hosts
PASS: run deploy host in multi-node deployment successfully

Story: 2010676
Task: 50244

Change-Id: I32c6cb08f56fa113e75db124652a52cc67b0d1a8
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>