Fix delete when failed to upload
If for some reason 'software upload <patch>' fails, 'software delete'
must be able to fully delete the load. In the case it tries to
delete a package from apt-ostree and does not find the component, it
means that the package does not exist, so it does not need to be
deleted.

Test plan:
PASS: Delete a patch after 'software upload' fails

Closes-bug: 2068926

Change-Id: I2b0559628182d6716e2aa0a15de57898e003fe67
Signed-off-by: Lindley Werner <Lindley.Vieira@windriver.com>