Run patch precheck in stx 9 environment
This commit is to allow running patch precheck in pre-USM
stx 9 release, where the legacy patch API is used for checking
whether the required patches are installed before the platform
upgrade.

The reason to use legacy patch API for the precheck is the USM
service doesn't exist in the stx 9 release.

Test Plan:

PASS: build and deploy iso
PASS: run the upgrade precheck in stx 9
PASS: run the upgrade precheck in the stx 10.

Task: 50338
Story: 2010676
Change-Id: I2b955b7db6db89676c418a331d610b1ababa1745
Signed-off-by: junfeng-li <junfeng.li@windriver.com>