Fix GA metadata release parsing
The parsing of the metadata was affected by commit [1], and
currently is returning a non-blocking error on the logs during
release upload and other operations.

This commit adds the missing tag on the metadata xml, and
does a small change on the metadata structure.

[1] https://review.opendev.org/c/starlingx/update/+/921700

Test Plan
PASS: build iso, verify metadata
PASS: upload iso, verify no errors on the logs
PASS: run software show and verify content and no errors

Change-Id: I9ea6dca4870fdaf5e9d3872c9cf66bc9ee4dcb74
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>