Fix script execution not being triggered
When installing a release, the script was not getting executed,
on software.log we couldn't see the script output.

This is happening because run-software-scripts bash script is called
as follows `run-software-scripts preinstall` (or postinstall).
The run-software-scripts looks for executable files within PATCH_SCRIPTDIR/$1
but PATCH_SCRIPTDIR wasn't defined anywhere so it couldn't find the scripts.

This fix declares the PATCH_SCRIPTDIR, so the scripts are found
and executed as expected.

The legacy implementation (sw-patch) has a very similar mechanism and the
PATCH_SCRIPTDIR properly declared
https://opendev.org/starlingx/update/src/commit/7847f7087e14f39f0e2e538b6cc91918e180bfd5/sw-patch/bin/patch-functions#L15

Another minor issue is that the run-software-scripts doesn't properly
report when there are no scripts, creating a false impression that
the script was executed.

Test Plan:
SUCCESS: preinstall and postinstall script execution is triggered properly.
SUCCESS: run-software-scripts <folder_name> properly logs when no scripts are present

Closes-bug: 2070391

Change-Id: I91f315d29171a19f9b5c8c09db7c40465dcb49b6
Signed-off-by: caio-volpato <caio.volpato@windriver.com>