Support for deploy scripts in RR patches
If a patch is reboot required, before reboot, it sets a flag:
agent_running_after_reboot_flag

When the host restarts, during the startup of the software agent, it
checks if this flag exists. If yes, it runs the post-install scripts.

TestPlan:
PASS: Deploy a RR patch with deploy scripts
PASS: Check in software.log if the scripts were executed

Story: 2010676
Task: 50591

Change-Id: Ie88d84849eda5f6c69779d99647c9836a7c9231b
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>