Fix deploy state sync after receiving ACK
There is a flaw in the active controller deploy state sync
when it sends his current deploy state to the standby controller,
then standby controller updates both json files software.json and
synced/software.json, but in the active controller side there is
not an update to the synced/software.json file. In this scenario,
a persistent alarm "900.231 Software deployment data is out of sync"
remains active in the system, blocking some actions, like swact, to
proceed with the major release deployment.

There is another cornercase that makes the alarm become stuck, that
is when the out-of-sync alarm is raised and an uncontrolled swact
occurs, the now-active controller won't be able to clear the alarm
even though the json files are synced, because the alarm_instance_id
used to create the alarm is the standby controller, that after the
swact assumes a different value.

This commit:
1. Fixes the bug, by updating synced/software.json file
   in the active controller side once it receives confirmation
   from the standby controller that it synced successfully
2. Fixes SoftwareMessageDeployStateUpdate by reverting [1]
   and instead only sending the message to controller nodes.
   The communication logic for deploy state update is migrated
   to software_controller now
3. Changes the out-of-sync alarm_instance_id assume generic
   "controller" value instead of the standby controller, to
   avoid the cornercase mentioned of when alarm is raised and
   a swact occurs, making alarm become stuck
4. Fixes minor tox issues introduced by other commits

[1] https://review.opendev.org/c/starlingx/update/+/921697

Test Plan
PASS: make the change in a system with the issue reproduced and
      observe the files being updated correctly and alarm goes
      away in the next audit check
PASS: (Standard 2+1) verify that the deploy state update message
      communication only occurs between controllers, by observing
      no errors on agent logs in computes
PASS: (Standard 2+1) full major release deploy

Closes-Bug: 2070525

Change-Id: I6ba1263f30366f756bb7d484370a2efe88632bc5
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>