Migrate etcd data to the folder
This commit is to migrate the etcd data to the TO release folder
in /opt/etcd/ during the platform upgrade.

Before this commit,  the migration was to set up a link  to the
FROM release folder in /opt/etcd/. This is problematic when the
FROM release folder is deleted when the upgrade is finished.

Test Plan:

PASS: built and deployed the iso on SX
PASS: etcd is running after deploy delete

Task: 50446
Story: 2010676
Change-Id: If5db91666003e73ccdb579943e16fbad858a34bb
Signed-off-by: junfeng-li <junfeng.li@windriver.com>