Fix metadata diverging from schema
During deploy start the commit id is added to the patch metadata, but
it was being added not following the metadata schema.

This review fix the way metadata tags are written and removed.
It also has a small fix to pass the correct state when changing it
to removing.

Test plan:
PASS: 'software deploy start' to remove a patch should succeed
PASS: Check 'software list' and 'software deploy show'

Story: 2010676
Task: 50573

Change-Id: If6e0470998687fd70fd6673c59da1e12aaf8227c
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>