Fix apply/remove order for multiple patches
When applying/removing one or more patches, a list of deployments
should be created with the correct order of deployment. This
commit fixes the algorithm to create this list.

It considers the cases where there is a single patch, multiple
patches, patches with sequential and non-sequential required lists.

Also fixes the deletion of ostree commit when removing a patch.

TestPlan adding:
PASS: 'software deploy start <one or more patches>' to apply
PASS: Check if the outputted list in the prompt is in order

TestPlan removing:
PASS: 'software deploy start <one or more patches>' to remove
PASS: Check if the outputted list in the prompt is in order

Story: 2010676
Task: 50681
Closes-bug: 2074237

Change-Id: Id9e0bcdf5a14f0946d8b7676dfecceecb01503de
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>