Fix set state to available after removing
At the end of an undeploy, after 'software deploy delete', the state
of the release stays on 'removing', but it should go to 'available'.
This commit fixes it.

TestPlan SX and DX:
PASS: Undeploy a release and check if the state is available
PASS: Deploy a major release and check if the state is deployed
      and 'No deploy in progress' after 'software deploy delete'

Story: 2010676
Task: 50699
Closes-bug: 2075187

Change-Id: Iba13e60998ff379528b45ce83ed655243671884c
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>