Fix: Remove changes verification in deploy host
A piece of code was added in 'deploy host' to check if a change in
the feed repo was notified to the PatchAgent during a patch
installation.
But this is causing crashes on some upgrade operations and a
verification was added into 'deploy start' to catch this error and
set the state to deploy-start-failed, avoiding it from reach
'deploy host'.

Fix on 'deploy start' side:
https://review.opendev.org/c/starlingx/update/+/924849

TestPlan:
PASS: 'Deploy host' a major release
PASS: The host deploy must not fail with 'Changes in feed not notified'

Closes-bug: 2075159

Change-Id: I2db488b8a2d5e3b027498eb7904401d4f5ca22c2
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>