Fix install scripts not found
When doing 'deploy host', it must copy and run the pre-install and
post-install scripts of all patches being applied.
But it was trying to find and run only the scripts of the last patch.

This commit fixes it. Instead of passing the last release ID to be
copied, it copies the scripts for all releases being applied.

TestPlan:
PASS: 'software deploy host' must run the scripts of all patches
      This can be checked in /var/log/software.log

Story: 2010676
Task: 50757

Change-Id: Id929b1926558501fc6ac94cb60ac38e41a4614c4
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>