Set deploy host fail after exception
This commit is to ensure when exception is thrown during the
deploy host preparation and before the request is sent to the
agent, the deploy host can be set to fail state. This will allow the
user to re-try the deploy host after exception is resolved.

Test Plan:

PASS: run unit tests
PASS: build and deploy the iso to DX
PASS: run the upgrade to deploy host

Task: 50802
Story: 2010676
Change-Id: Iddf1b7ca7e34f891e0e59e7e9c3be7e3e13c099c
Signed-off-by: junfeng-li <junfeng.li@windriver.com>