Send cleanup request to agents only on success deployment
Currently, the cleanup request message is sent to agents during
deploy delete, regardless of the current deployment state. This
behavior is incorrect because if the deployment is deleted in
start-done or start-failed states the agents will update their
ostree remotes incorrectly pointing to the to-release feed.

This commit fixes this behavior by sending the cleanup message
to the agents only if deployment completed or was rolled-back
successfully.

Test Plan
PASS: AIO-DX - deploy delete on start-[done|failed] states,
      verify the agents do not update the ostree remotes
PASS: AIO-DX - deploy delete on completed/host-rollback-done
      states, verify the agents update the ostree remotes and
      remove the flags correctly

Story: 2010676
Task:

Change-Id: Icf00c59a85d2bbec4ca8d864c42be5dd4a8c4919
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>