software-agent: properly query remote feed
Ostree does not support using a url on the ostree log command and
the feed repo is not acessible from non-controller nodes. This is
a problem if we are trying to determine if the local repo is in sync
with the feed on compute/storage nodes.

This commit adds a new "controller-feed" remote to non-controller
nodes that points to the feed on the active controller. software-agent
can use this remote to log the latest commit.

Also modifies the gpg-verification workaround to make sure that
gpg verification is disabled.

Test-Plan
PASS: Standard Install
PASS: Compute nodes able to query feed
PASS: Lock/Unlock worker nodes after host-swact
PASS: deploy major patch on worker nodes

Story: 2010676
Task: 50879

Change-Id: I5153115d9d1ae88ff322d9c5179483229a5378f8
Signed-off-by: mmachado <mmachado@windriver.com>