Adjust k8s symlink remove hook to run only in controllers
This commit limits the scope of the hook to the controllers,
which is the type of host where the symlink is expected to
exist. Without this commit, rollback fails in other types of
host.

Test Plan
PASS: run host-rollback in controllers
PASS: run host-rollback in other types of host

Story: 2010676
Task: 50896

Change-Id: I96b9415a53118c81990bb7c1c3a28ee4d2d38e50
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>