Change USM internal backend port
To implement load balance in using haproxy in internal endpoints, a
new proxy frontend was included and binded to the mgmt IP and the
private port.

Considering this, the default backend port had to be changed from the
private port. Since in DC the admin frontend also uses the mgmt IP and
(private port + 1), the default backend was changed to (private port +
2), and the alternative backend to the next value.

The ports were changed from -> to:
- backend default:     5497 -> 5499
- alt backend (slow):  5499 -> 5500

Test plan:
PASS: Bootstrap and unlock AIO-SX, verify haproxy.cfg to see if the
      new frontend is in place and the backend ports are correct.
PASS: Bootstrap and unlock DC + SX subcloud, verify haproxy.cfg to
      see if the new frontend is in place and the backend ports are
      correct.
PASS: Execute 'software deploy precheck', verify that the 'precheck'
      request is reaching the alternative backend.
PASS: Execute 'sw-manager sw-deploy-strategy create', verify that the
      'precheck' request is reaching the alternative backend.

Story: 2010676
Task: 50912

Change-Id: I234c1db24c6a24c052fbb5933d4804c0a4ed139b
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>