Improve removal of additional ceph-mon on USM rollback
When upgrading to stx.10 and doing a rollback, Ceph is becoming
unresponsive, breaking the host-rollback action.

The fixed Ceph monitors from both controllers should be removed from
the monmap when rolling back the deployment. Currently, only the
active controller's local fixed monitor is being removed leaving the
standby's fixed monitor in the monmap. When the standby controller
was in 'from-version' the active one was in 'to-version' and a
host-swact is triggered, Ceph mon.controller will start in the new
active controller and detect the other fixed monitor running on the
standby controller. Then there will be 2 monitors running and Ceph
needs an odd number of monitors to form a quorum, leading to Ceph
stopping responding.

Ceph can run with only one or two monitors when it is getting set up,
but as soon as one monitor is disabled and started back, Ceph will
recognize the setup is done and will begin normal operation, requiring
the correct number of monitors running to form a quorum.

Removing both fixed monitors when running host-rollback command
ensures there will only be the mon.controller in the monmap and Ceph
will always be responsive.

Test-Plan:
  PASS: AIO-DX: Start upgrade to stx.10; deploy one controller; abort
  the upgrade; host-rollback the upgraded controller; no errors should
  be found.

  PASS: AIO-DX: Start upgrade to stx.10; deploy both controllers;
  abort the upgrade; host-rollback the upgraded controllers; no errors
  should be found.

Story: 2010676
Task: 50892

Change-Id: If8bf53fae1a9d1f5e264d22394cf751ae6795059
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>