Ensure software controller daemon exit in error
When a thread of software controller daemon fails (dead), the entire
daemon should exit, and pmon will then restart it with raising proper
alarm as it is configured.

This change updates the logic between threads, to ensure that all
threads are informed and gracefully stopped after a failure occurs in a
thread.

TCs:
    passed: (simulated) trigger uncaught exception in all 4 threads of
            the daemon, observed the daemon exits soon after and pmond
            restarts software controller daemon.

Story: 2010676
Task: 50916

Change-Id: I0489aea2f7bee953605cfb49d3f041b166d65fde
Signed-off-by: Bin Qian <bin.qian@windriver.com>