Clean files if upload fails
If a 'software upload <patch>' fails, it notifies the user but keeps
the corrupted files in the system memory.
This commit fixes it by, in case of failure, cleaning all files from
the release being uploaded.

Test-Plan:
PASS: Upload and deploy a patch with success
Upload a corrupted patch:
PASS: See the error message in the prompt
PASS: See in the logs the initial uploaded files are now deleted

Depends-on:
https://review.opendev.org/c/starlingx/apt-ostree/+/929535

Story: 2010676
Task: 50944

Change-Id: I667acd47ad1c8f1bd19e4661f4c18ed7b189bdc3
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>