Rollback patch during apply
This commit adds the capacity to abort and rollback the nodes if the
deployment fails in one or multiple of them. If the user decides to
abort a deployment, it will remove all the associated commits from
the feed repo and notify all hosts. Then during host-rollback the
nodes will sync up with the feed roll back.

After the rollback, the user must run "software deploy delete" to
clean the deployment.

Test Plan SX and DX:
PASS: Rollback while removing is not allowed
PASS: Rollback with success an inservice patch
PASS: Rollback with success an RR patch
PASS: Rollback a major release deployment with success

Story: 2010676
Task: 50931

Change-Id: I881a8134f52cb566b92c4204ee3e76bfa49b508c
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>