Create new endpoint to return if host is next to be deployed.
This commit add's a new endpoint which receives the hostname as
parameter and returns True or False to indicate if the given hostname
is the next to be deployed.

Test Plan:
PASS: Lock/Unlock controller-1 after software deploy abort with both
controllers in deploy-host-rollback-pending state.
PASS: Check the endpoint return value is accurate.

Story: 2010676
Task: 50980

Change-Id: I365ad2c4ee622ee5cc191c75991d5402b721e1a2
Signed-off-by: Luis Eduardo Bonatti <LuizEduardo.Bonatti@windriver.com>