Fix import script param parsing bool value
The issue to fix is when the default boolean value True is passed
to the is_usm_iso, it is not validated as the condition check is
to only use string 'True'.

This commit is to ensure when the default boolean value is passed to the
usm_load_import script,  it can be parsed as the boolean script.

Test Plan:

PASS: build the iso
PASS: upload the iso to 22.12 system
PASS: upload N-1 load on system control that contains this fix.

Task: 50993
Story: 2010676
Change-Id: I1d7651d0ff26d31e52589f1b039a788c058fdb3d
Signed-off-by: junfeng-li <junfeng.li@windriver.com>