Sync inactive load to controller-1
This commit is to sync the N-1 feed to controller-1 after the
load is uploaded on the system controller.  This change will allow
controller-1 to support subcloud N-1 load deployment after the swact.

Test Plan:

PASS: build and deploy the iso on system controller
PASS: upload N-1 load and verify the feed is sync'ed on controller-1
PASS: upload N+1 load

Task: 51056
Story: 2010676

Change-Id: Ia1866c46d35bb94348d8805a38534a222513639a
Signed-off-by: junfeng-li <junfeng.li@windriver.com>