Check before deleting old deployments
When the Software Agent starts, it lists all ostree deployments, and
deletes all with an index greater than 2 in the list. But it does not
check the list size before trying to delete, leading in some cases to
an out-of-range exception.

This commit adds a check before iterating through the list to delete.

Test-Plan:
PASS: Deploy an ISO and check no out-of-range error after unlock
PASS: Deploy an RR patch and see only two deployments after unlock

Story: 2010676
Task: 51069

Change-Id: I2555597fac30677bcb36629a2ea66feb52a40de1
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>