USM patch reapply fails after wrong release order
This commit fixes an issue where the the patch release order validation
raises an exception after the update in software.json file, causing it
to be saved in an invalid state and blocking further patches from being
applied.

Test plan:
All of the tests were performed using an overlay of the
system_controller.py file and restarting all of the sofware services
with systemctl.
1. PASS: Upload a P2 that requires a P1, apply the software strategy and
   validate that an error occurs, displaying the message "Software
   deploy start was rejected: Not all required releases are uploaded,
   missing <P1 patch>"
2. PASS: Verify that the software.json file was not created.
3. Delete the strategy, apply the P1 patch, recreate the strategy for
   the P2 and apply it, confirming that it finishes successfully.

Closes-Bug: 2082839

Change-Id: I98cb171a0f74f3cbfda5c13b17651361143b211b
Signed-off-by: Raphael Luiz Avelar Lima <Raphael.Lima@windriver.com>