Improve require_deploy_state for complete and deploy start
This commit improves the error return message for deploy complete and
deploy start. If deploy is in start-failed and the user try to start
again the return message is a bit weird as the action needed is delete.
For deploy complete message is to be more precise in scenarios like two
consecutive attempts, even with the deployment already completed.

Test Plan:
PASS: 2 deploy complete in a row and check the message.
PASS: deploy complete with start-done and check the message.
PASS: deploy start <release> with start and start-failed state.

Story: 2010676
Task: 51093

Change-Id: I6f1b895f8a3c2adc634aeeef1dcdd92e0947ff94
Signed-off-by: Luis Eduardo Bonatti <luizeduardo.bonatti@windriver.com>