Fix N-1 release deletion in system controller
A release, when deleted from the system controller, has its files in the
unavailable and committed directories from /opt/software deleted before
the metadata file removal. Because of that, when removing a N-1, the
metadata file is removed before the step that performs its clean up,
causing a MetadataFail exception.
In this commit, a validation is added to prevent an exception from being
raised raised when the file was already deleted.

Test plan:
All test cases were executed by modifying the software_controller.py in
a running system and restarting all software services with systemctl.
1. PASS: Upload a N-1 release using --local parameter and verify that it
   finishes successfully.
2. PASS: Delete the N-1 release with the software delete command and
   verify that a success message is displayed and the appropriate
   warning is logged to software.log.
3. PASS: Verify that the N-1 release directory in
   /opt/dc-vault/playbooks and /opt/dc-vault/software is removed.
4. PASS: Verify that the N-1 release files in
   /opt/software/metadata/unavailable and
   /opt/software/metadata/committed are removed.

Closes-bug: 2083647

Change-Id: Ib5e173920ecb69d48709f6e836c5d793db6ee0e8
Signed-off-by: Raphael Luiz Avelar Lima <Raphael.Lima@windriver.com>