Improve feed repo sync script to detect inactive peer
This commit allows detecting the inactive node in a duplex subcloud
to which the changes of the active node must be synchronized.

When the content of the feed repo directory on the active node is
modified, it must be synchronized with the inactive node.

This only applies to a duplex system. On a simplex subcloud, if the
script is invoked, the synchronization will simply be skipped.

Test Plan:
PASS - Bring up a 24.09 system controller.
     - Bring up a 24.09 duplex subcloud.
     - Upload and deploy patch 24.09.1 on subcloud.
     - Verify in software.log that the ostree feed repo is
       synchronized and check the directory content.

PASS - Bring up a 24.09 system controller.
     - Bring up a 24.09 simplex subcloud.
     - Upload and deploy patch 24.09.1 on subcloud.
     - Verify in software.log that the synchronization is
       skipped.

Story: 2010676
Task: 51147

Change-Id: I94a3f0f78b08368ebfb6f7dba82b4fab245580f7
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>