Exclude patching ops from major release deploy
In a pre-patched iso, the metadata file comes with the latest patch may
contain patching specific pre and post install scripts. However the
scripts are not accessible as part of the pre-patched iso.

This change is to exclude copying pre-install and post-install scripts
for patching deploy from being applied to major release deploy.

Story: 2010676
Task: 51157

TCs:
    passed: complete SX upgrade to a pre-patched iso

Change-Id: I0fb6378609c6f371483c4fd8547e3eda202bc65b
Signed-off-by: Bin Qian <bin.qian@windriver.com>