Remove to-release remote and ref if deploy host fails
Currently if deploy host fails after the remote create/pull but
before to-release ostree commit-id is deployed in fact then the
system will incorrectly consider that to-release commit-id is
deployed.

This commit fixes the issue. The cause is that when the remote is
pulled it created a 'starlingx' ref with to-release commit-id on
sysroot. The fix is, in case of failure, remove both the remote
and the ref, so that system is returned to it's original status
before deploy host.

Test Plan
PASS: force deploy host failure after remote create/pull, verify
      to-release remote and ref are deleted
PASS: force deploy host failure on pre-hooks execution, verify
      to-release remote and ref are deleted
PASS: on both previous TC, retry deploy host and verify it runs
      as expected

Story: 2010676
Task: 51172

Change-Id: I302faa57bb677148992adb4f965b512534037e42
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>