Change pxeboot files rsync source location
For subclouds, where prestage is used, in 'deploy host' stage
an agent hook tries to copy to-release pxe files from the to-release
feed, but they do not exist. In standalone clouds this issue wasn't
seen as 'software upload' copies the files earlier in the deployment
process, but in the prestaged process 'software upload' does not run.

This commit fixes the issue by copying the pxe files directly from
the to-release ostree repo once it is deployed.

It also fixes a permission issue on pxeboot-update-xx.yy.sh file,
because usm_load_import changes it to 755. This was blocking the
script from executing, hence blocking host-unlock.

Some comments and docstrings were reworded for clarity.

Test Plan
PASS: on an AIO-SX prestaged subcloud, apply the commit manually
      and run 'deploy host' -> host-unlock successfully
PASS: (Regression) AIO-DX deploy host successfully

Story: 2010676
Task: 51180

Change-Id: Icba72e9951e903685d162334232077aafced341e
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>