Allow upload local flag run on remote terminal
This commit changes --local flag behavior in 'software upload'.
After the change, --local flag can be set on the remote terminal.

When the flag is set, the 'software upload' execution will look for
the specified files on the active controller. The file extension
validation is done on client side, while the file existence validation
is run on the server side.

Test Plans:
PASS: build and deploy iso on SX
PASS: upload files that are existing on server with --local set
PASS: upload files that are not existing on the server with --local set
PASS: upload files without --local set

Story: 2010676
Task: 51183

Change-Id: Ibaf4c0c9dc8fdab025067bc9fa97e1839323ac01
Signed-off-by: junfeng-li <junfeng.li@windriver.com>