Avoid reboot loop when patch fails
When there is a fail during remote pull a reboot-required patch to a
host, the system enters a reboot loop mode. This happens because the
reboot required flag was left set even with a failure in the
installation process.

This commit fixes it by checking if the flag still exists after no
changes were made by the install, it deletes the flag and returns a
proper error.

Test-plan:
PASS: Fail a patch during deploy host
      - The system should set the state to host-failed
      - The system must not enter in reboot loop

Story: 2010676
Task: 51192

Change-Id: Ib311f28911620cd14df357e06ff9e5afcf82b745
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>