Upload, move and delete activate scripts
This commit allows the code to extract the patch activation scripts
from the patch during the upload, it saves the scripts under
/opt/software/software-scripts folder with the following name
pattern: releaseId_scriptName.

During the deploy start, the script of the deploying or removing
patches are copied to /etc/update.d, the folder is created and the
script is saved using only its name scriptName. If multiple patches
are being deployed and they have scripts with the same name, only one
copy of the script will be saved (to avoid doing the same operation
multiple times).

During deploy delete, the folder /etc/update.d is deleted.

During software delete, the scripts of being deleted patch are
removed from /opt/software/software-scripts.

Test-Plan:
PASS: Upload a patch and see the scripts under the
      software-scripts folder, following the name pattern.
PASS: Deploy start to apply a patch and see the update.d folder was
      created and the scripts copied to it.
PASS: Deploy start to remove a patch and see the update.d folder was
      created and the scripts copied to it.
PASS: Deploy delete and see if the update.d folder was deleted
PASS: Software delete a patch and see if the scripts were deleted
      from software-scripts.
PASS: Upload a patch with activate scripts on metadata, but missing
      the actual script and see if the upload failed

Story: 2010676
Task: 51153

Change-Id: Iff33d2d94b69e8ae4479259ffd00fa708d5de518
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>